kern: Remap the framebuffer at a high virtual address
This commit actually contains multiple changes.

* In the linker script all the __<section>_{start,end} symbols have
been aligned to 4 bytes. This was the source of a bug where all
externally defined symbols would be read as zero after the BSS
was cleared. I'm still unsure about this fix but it seems to be
working
* The whole .prekernel.data section has been removed and merged
in the normal .data section. The prekernel now only consists of
that tiny code stub in start.S
* Many references to the old peripherals addresses have been updated
* A new header "areas" with definitions for the higher half areas has
been added